Update isFieldEnabled to handle both schema types
