+++ front/app/containers/IdeasShow/index.tsx
@@ -33,7 +33,7 @@ import MobileSharingButtonComponent from './Buttons/MobileSharingButtonComponent
 import RightColumnDesktop from './RightColumnDesktop';
 
 // utils
-import { checkFieldEnabled } from './isFieldEnabled';
+import { isFieldEnabled } from 'utils/projectUtils';
 
 // resources
 import GetIdeaImages, {
@@ -218,12 +218,6 @@ export const IdeasShow = ({
   const ideaflowSocialSharingIsEnabled = useFeatureFlag({
     name: 'ideaflow_social_sharing',
   });
-  const isIdeaCustomFieldsEnabled = useFeatureFlag({
-    name: 'idea_custom_fields',
-  });
-  const isDynamicIdeaFormEnabled = useFeatureFlag({
-    name: 'dynamic_idea_form',
-  });
 
   const ideaCustomFieldsSchemas = useIdeaCustomFieldsSchemas({
     projectId,
@@ -276,12 +270,10 @@ export const IdeasShow = ({
       compact === true ||
       (windowSize ? windowSize <= viewportWidths.tablet : false);
 
-    const proposedBudgetEnabled = checkFieldEnabled(
+    const proposedBudgetEnabled = isFieldEnabled(
       'proposed_budget',
       ideaCustomFieldsSchemas,
-      locale,
-      isIdeaCustomFieldsEnabled,
-      isDynamicIdeaFormEnabled
+      locale
     );
 
     content = (
